+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Add configurable manifest.yaml.template path for build helper
 
+### Changed
+
+- Bump Rust toolchain from 1.85 to 1.93
+
 ## [3.1.0]
 
 ### Added